refactor: change i18n to localization
