+++ components.json
@@ -12,7 +12,7 @@
   },
   "aliases": {
     "components": "@/app/components",
-    "utils": "@/lib/utils",
+    "utils": "@/app/utils",
     "ui": "@/app/components/ui",
     "lib": "@/lib",
     "hooks": "@/app/hooks"
