+++ middleware.ts
@@ -1,5 +1,5 @@
 import createMiddleware from "next-intl/middleware";
-import { routing } from "./app/i18n/routing";
+import { routing } from "./app/features/localization/routing";
 
 export default createMiddleware(routing);
